change float to double

+++ src/Lib.hs
@@ -13,40 +13,40 @@ import qualified Text.PrettyPrint.Tabulate as T
 
 data RangeInfo = RangeInfo
     { time_     :: Integer
-    , angle_    :: Float
-    , velocity_ :: Float
-    , range_    :: Float
+    , angle_    :: Double
+    , velocity_ :: Double
+    , range_    :: Double
     } deriving (Data, G.Generic, Show)
 
 -- | Acceleration of gravity
-g :: Float
+g :: Double
 g = 9.80665
 
 -- | Converts from degrees to radians.
-radians :: Float -> Float
+radians :: Double -> Double
 radians degrees = degrees * (pi / 180)
 
 -- | Calculates the sine of an angle in degrees.
-sin' :: Float -> Float
+sin' :: Double -> Double
 sin' degrees = sin $ radians degrees
 
 -- | Calculates the cosine of an angle in degrees.
-cos' :: Float -> Float
+cos' :: Double -> Double
 cos' degrees = cos $ radians degrees
 
-timeOfFlight :: Float -> Float -> Float -> Float
+timeOfFlight :: Double -> Double -> Double -> Double
 timeOfFlight angle velocity range = range / (velocity * cos' angle)
 
-maxHorizontalDistance :: Float -> Float -> Float
+maxHorizontalDistance :: Double -> Double -> Double
 maxHorizontalDistance angle velocity = velocity ^ 2 * sin' (2 * angle) / g
 
-horizontalDistance :: Float -> Float -> Integer -> Float
+horizontalDistance :: Double -> Double -> Integer -> Double
 horizontalDistance angle velocity time = do
     let maxDistance = maxHorizontalDistance angle velocity
     let distance = velocity * fromInteger time * cos' angle
     min maxDistance distance
 
-rangeInfo :: Float -> Float -> Integer -> RangeInfo
+rangeInfo :: Double -> Double -> Integer -> RangeInfo
 rangeInfo angle velocity time = do
     let shellRange = horizontalDistance angle velocity time
     RangeInfo {
